Jenkinsfile

+++ Jenkinsfile
@@ -1,10 +1,9 @@
 pipeline {
     agent none
     parameters {
-        //string(name: 'BC_LIB', defaultValue: '', description: 'braincloud-unreal-plugin-src branch (blank for .gitmodules)')
-        string(name: 'BRANCH_NAME', defaultValue: 'jenkins-build', description: 'examples-unreal branch')
-        //choice(name: 'PROJECTNAME', choices: ['all', 'RelayTestApp', 'RelayTestAppCPP', 'TappyChicken', 'ScriptTestApp','Groups', 'Leaderboard'], description: 'Where to run tests?')
-        //choice(name: 'PLATFORM', choices: ['all', 'Mac', 'Win64', 'IOS', 'Android'])
+        string(name: 'BC_LIB', defaultValue: '', description: 'braincloud-unreal-plugin-src branch (blank for .gitmodules)')
+        string(name: 'BRANCH_NAME', defaultValue: 'develop', description: 'examples-unreal branch')
+        booleanParam(name: 'CLEAN_BUILD', defaultValue: true, description: 'clean pull and build')
     }
     stages {
 
@@ -20,20 +19,21 @@ pipeline {
                 BRAINCLOUD_TOOLS="/Users/buildmaster/braincloud-client-master"
             }
             steps {
-                deleteDir()
                 echo "---- braincloud Code Pull ${BRANCH_NAME} ${BC_LIB}"
-                checkout([$class: 'GitSCM', branches: [[name: '*/${BRANCH_NAME}']], extensions: [[$class: 'SubmoduleOption', disableSubmodules: false, parentCredentials: false, recursiveSubmodules: true, reference: '', trackingSubmodules: false]], userRemoteConfigs: [[url: 'https://github.com/getbraincloud/examples-unreal.git']]])
-                sh 'autobuild/checkout-submodule.sh ${BC_LIB}' //not ported to windows yet
+                if (${params.CLEAN_BUILD}) {
+                    deleteDir()
+                }
+                checkout([$class: 'GitSCM', branches: [[name: '*/${BRANCH_NAME}']], extensions: [[$class: 'SubmoduleOption', disableSubmodules: false, parentCredentials: false, recursiveSubmodules: true, reference: '', trackingSubmodules: false]], userRemoteConfigs: [[url: 'https://github.com/getbraincloud/examples-unreal.git']]])				
+                sh 'autobuild/checkout-submodule.sh ${BC_LIB}'
                 sh 'autobuild/_brainCloudSetup_examples-unreal.command'
                 sh 'autobuild/makebuild.sh RelayTestApp MAC UE_5_Mac'
-                sh 'autobuild/makebuild.sh RelayTestAppCpp MAC UE_5_Mac'
+                //sh 'autobuild/makebuild.sh RelayTestApp IOS'
+                //sh 'autobuild/makebuild.sh RelayTestApp ANDROID'
+/*                 sh 'autobuild/makebuild.sh RelayTestAppCpp MAC UE_5_Mac'
                 sh 'autobuild/makebuild.sh TappyChicken MAC UE_5_Mac'
                 sh 'autobuild/makebuild.sh ScriptTestApp MAC UE_5_Mac'
                 sh 'autobuild/makebuild.sh Groups MAC UE_5_Mac'
-                sh 'autobuild/makebuild.sh Leaderboard MAC UE_5_Mac'
-                //sh 'autobuild/makebuild.sh RelayTestApp IOS'
-                //sh 'autobuild/makebuild.sh RelayTestApp ANDROID'
-                //sh 'autobuild/makebuild.sh ${params.PROJECTNAME} ${params.PLATFORM} UE_5_${params.PLATFORM}'
+                sh 'autobuild/makebuild.sh Leaderboard MAC UE_5_Mac' */
             }
             post {
                 success {
@@ -53,22 +53,23 @@ pipeline {
                  BRAINCLOUD_TOOLS="C:\\Users\\buildmaster\\braincloud-client-master"
               }
              steps {
-                echo "---- braincloud Code Pull ${BRANCH_NAME} ${BC_LIB}"
-                deleteDir()
-                checkout([$class: 'GitSCM', branches: [[name: '*/${BRANCH_NAME}']], extensions: [[$class: 'SubmoduleOption', disableSubmodules: false, parentCredentials: false, recursiveSubmodules: true, reference: '', trackingSubmodules: false]], userRemoteConfigs: [[url: 'https://github.com/getbraincloud/examples-unreal.git']]])
-                bat 'autobuild\\_brainCloudSetup_examples-unreal.bat'
-                bat 'autobuild\\makebuild.bat RelayTestApp Win64 UE_5_Win64'
-                bat 'autobuild\\makebuild.bat RelayTestAppCPP Win64 UE_5_Win64'
-                bat 'autobuild\\makebuild.bat TappyChicken Win64 UE_5_Win64'
-                bat 'autobuild\\makebuild.bat ScriptTestApp Win64 UE_5_Win64'
-                bat 'autobuild\\makebuild.bat Groups Win64 UE_5_Win64'
-                bat 'autobuild\\makebuild.bat Leaderboard Win64 UE_5_Win64'
-                //bat 'autobuild\\makebuild.bat ${params.PROJECTNAME} ${params.PLATFORM} UE_5_${params.PLATFORM}'
+                 echo "---- braincloud Code Pull ${BRANCH_NAME} ${BC_LIB}"
+                 //if (${CLEAN_BUILD}) {
+                    deleteDir()
+                 //}
+                 checkout([$class: 'GitSCM', branches: [[name: '*/${BRANCH_NAME}']], extensions: [[$class: 'SubmoduleOption', disableSubmodules: false, parentCredentials: false, recursiveSubmodules: true, reference: '', trackingSubmodules: false]], userRemoteConfigs: [[url: 'https://github.com/getbraincloud/examples-unreal.git']]])
+                 bat 'autobuild\\_brainCloudSetup_examples-unreal.bat'
+                 bat 'autobuild\\makebuild.bat RelayTestApp Win64 UE_5_Win'
+                 //bat 'autobuild\\makebuild.bat RelayTestAppCPP Win64 UE_4_Win'
+                 //bat 'autobuild\\makebuild.bat TappyChicken Win64 UE_4_Win'
+                 //b/at 'autobuild\\makebuild.bat ScriptTestApp Win64 UE_4_Win'
+                 //bat 'autobuild\\makebuild.bat Groups Win64 UE_4_Win'
+                 //bat 'autobuild\\makebuild.bat Leaderboard Win64 UE_4_Win'
             }
             post {
                 success {
-                    fileOperations([fileZipOperation(folderPath: 'UE_5_Win64', outputFolderPath: '.')])
-                    archiveArtifacts allowEmptyArchive: true, artifacts: 'UE_5_Win64.zip', followSymlinks: false, onlyIfSuccessful: true
+                    fileOperations([fileZipOperation(folderPath: 'UE_5_Win', outputFolderPath: '.')])
+                    archiveArtifacts allowEmptyArchive: true, artifacts: 'UE_5_Win.zip', followSymlinks: false, onlyIfSuccessful: true
                 }
             }
        }
@@ -84,19 +85,20 @@ pipeline {
               }
              steps {
                  echo "---- braincloud Code Pull ue4-examples ${BC_LIB}"
-                 deleteDir()
+                 //if (${CLEAN_BUILD}) {
+                     deleteDir()
+                 //}
                  checkout([$class: 'GitSCM', branches: [[name: '*/ue4-examples']], extensions: [[$class: 'SubmoduleOption', disableSubmodules: false, parentCredentials: false, recursiveSubmodules: true, reference: '', trackingSubmodules: false]], userRemoteConfigs: [[url: 'https://github.com/getbraincloud/examples-unreal.git']]])
                  bat 'autobuild\\_brainCloudSetup_examples-unreal-4.bat'
-                 bat 'autobuild\\makebuild.bat RelayTestApp Win64 UE_4_Win64'
-                 bat 'autobuild\\makebuild.bat RelayTestAppCPP Win64 UE_4_Wi64'
-                 bat 'autobuild\\makebuild.bat TappyChicken Win64 UE_4_Win64'
-                 bat 'autobuild\\makebuild.bat ScriptTestApp Win64 UE_4_Win64'
-                 //bat 'autobuild\\makebuild.bat ${params.PROJECTNAME} ${params.PLATFORM} UE_4_${params.PLATFORM}'
+                 bat 'autobuild\\makebuild.bat RelayTestApp Win64 UE_4_Win'
+                // bat 'autobuild\\makebuild.bat RelayTestAppCPP Win64 UE_4_Win'
+                 //bat 'autobuild\\makebuild.bat TappyChicken Win64 UE_4_Win'
+                 //bat 'autobuild\\makebuild.bat ScriptTestApp Win64 UE_4_Win'
              }
             post {
                 success {
-                    fileOperations([fileZipOperation(folderPath: 'UE_4_Win64', outputFolderPath: '.')])
-                    archiveArtifacts allowEmptyArchive: true, artifacts: 'UE_4_Win64.zip', followSymlinks: false, onlyIfSuccessful: true
+                    fileOperations([fileZipOperation(folderPath: 'UE_4_Win', outputFolderPath: '.')])
+                    archiveArtifacts allowEmptyArchive: true, artifacts: 'UE_4_Win.zip', followSymlinks: false, onlyIfSuccessful: true
                 }
             }
        }
